chore: move download components to own directory
